fix evd and sequential workflow
